--- support/reach/src/api/cc/mod.rs
@@ -1,6 +1,0 @@
-mod message;
-pub use message::ReachMessageCcApi;
-mod trigger_scene;
-pub use trigger_scene::ReachTriggerSceneCcApi;
-
-pub type ReachCcApi = (ReachTriggerSceneCcApi, ReachMessageCcApi);
