--- support/reach/src/api/ci/mod.rs
@@ -1,2 +1,0 @@
-mod message;
-pub use message::ReachMessageCiApi;
