+++ support/reach/src/api/reach_api_cc/reach_api_cc_message.rs
@@ -9,13 +9,13 @@ use tardis::web::poem_openapi;
 use tardis::web::poem_openapi::payload::Json;
 use tardis::web::web_resp::{TardisApiResult, TardisResp, Void};
 
-use crate::client::{GenericTemplate, SendChannelMap};
-use crate::config::ReachConfig;
-use crate::consts::*;
+use crate::reach_send_channel::{GenericTemplate, SendChannelMap};
+use crate::reach_config::ReachConfig;
+use crate::reach_consts::*;
 use crate::dto::*;
-use crate::init::REACH_SEND_CHANNEL_MAP;
+use crate::reach_init::REACH_SEND_CHANNEL_MAP;
 #[cfg(feature = "simple-client")]
-use crate::invoke::Client;
+use crate::reach_invoke::Client;
 use crate::serv::*;
 
 #[derive(Clone)]
