+++ support/reach/src/api/reach_api_ct/reach_api_ct_msg_log.rs
@@ -1,7 +1,7 @@
-use crate::consts::get_tardis_inst;
+use crate::reach_consts::get_tardis_inst;
 use crate::dto::*;
 #[cfg(feature = "simple-client")]
-use crate::invoke::Client;
+use crate::reach_invoke::Client;
 use crate::serv::*;
 use bios_basic::rbum::serv::rbum_crud_serv::RbumCrudOperation;
 use bios_sdk_invoke::simple_invoke_client;
