+++ support/reach/src/api/reach_api_ct/reach_api_ct_msg_template.rs
@@ -7,12 +7,12 @@ use tardis::web::poem_openapi::payload::Json;
 use tardis::web::poem_openapi;
 use tardis::web::web_resp::{TardisApiResult, TardisPage, TardisResp};
 
-use crate::consts::get_tardis_inst;
+use crate::reach_consts::get_tardis_inst;
 use crate::dto::*;
 use crate::serv::*;
 
 #[cfg(feature = "simple-client")]
-use crate::invoke::Client;
+use crate::reach_invoke::Client;
 
 use super::map_notfound_to_false;
 #[derive(Clone, Default)]
