+++ support/reach/src/lib.rs
@@ -1,19 +1,18 @@
 #![warn(clippy::dbg_macro)]
 #![warn(clippy::unwrap_used)]
 
-pub mod api;
-pub mod client;
-pub mod config;
-pub mod consts;
+mod api;
+pub mod reach_send_channel;
+pub mod reach_config;
+pub mod reach_consts;
 mod domain;
 pub mod dto;
-mod errors;
-mod init;
+mod reach_init;
 
 #[cfg(feature = "simple-client")]
-pub mod invoke;
+pub mod reach_invoke;
 mod task;
-pub use init::init;
+pub use reach_init::init;
 pub mod reach_initializer {
     pub use crate::init;
 }
