+++ support/reach/src/reach_init.rs
@@ -14,9 +14,9 @@ use tardis::{
 
 use crate::{
     api,
-    client::SendChannelMap,
-    config::ReachConfig,
-    consts::{get_tardis_inst, DOMAIN_CODE, DOMAIN_REACH_ID, RBUM_EXT_TABLE_REACH_MESSAGE, RBUM_KIND_CODE_REACH_MESSAGE, REACH_INIT_OWNER},
+    reach_send_channel::SendChannelMap,
+    reach_config::ReachConfig,
+    reach_consts::{get_tardis_inst, DOMAIN_CODE, DOMAIN_REACH_ID, RBUM_EXT_TABLE_REACH_MESSAGE, RBUM_KIND_CODE_REACH_MESSAGE, REACH_INIT_OWNER},
     serv::ReachTriggerSceneService,
     task,
 };
