+++ support/reach/tests/test_message_send.rs
@@ -2,7 +2,7 @@ use bios_basic::rbum::dto::rbum_item_dto::RbumItemAddReq;
 use tardis::{basic::result::TardisResult, log, serde_json::json, testcontainers, tokio};
 
 mod test_reach_common;
-use bios_reach::{consts::*, dto::*, invoke};
+use bios_reach::{reach_consts::*, dto::*, reach_invoke};
 use test_reach_common::*;
 #[tokio::test(flavor = "multi_thread")]
 pub async fn test_ct_api() -> TardisResult<()> {
@@ -14,7 +14,7 @@ pub async fn test_ct_api() -> TardisResult<()> {
     let holder = init_tardis(&docker).await?;
     let ctx = get_test_ctx();
     let funs = get_tardis_inst();
-    let client = invoke::Client::new("http://localhost:8080/reach", ctx, &funs);
+    let client = reach_invoke::Client::new("http://localhost:8080/reach", ctx, &funs);
     const CONTENT_TEMPLATE: &str = "hello {name}, your code is {code}";
     let template_name = random_string(16);
     fn expected_content(name: &str, code: &str) -> String {
