+++ support/reach/tests/test_send_client.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 use tardis::{basic::result::TardisResult, testcontainers, tokio};
 
 mod test_reach_common;
-use bios_reach::{consts::*, invoke};
+use bios_reach::{reach_consts::*, reach_invoke};
 use test_reach_common::*;
 #[derive(Deserialize)]
 pub struct TestConfig {
@@ -30,7 +30,7 @@ pub async fn test_hw_sms() -> TardisResult<()> {
     let holder = init_tardis(&docker).await?;
     let ctx = get_test_ctx();
     let funs = get_tardis_inst();
-    let client = invoke::Client::new("http://localhost:8080/reach", ctx, &funs);
+    let client = reach_invoke::Client::new("http://localhost:8080/reach", ctx, &funs);
     // client.pwd_send(&phone, &code, &()).await?;
 
     client.vcode_send(&phone, &code, &()).await?;
@@ -50,7 +50,7 @@ pub async fn test_mail() -> TardisResult<()> {
     let holder = init_tardis(&docker).await?;
     let ctx = get_test_ctx();
     let funs = get_tardis_inst();
-    let client = invoke::Client::new("http://localhost:8080/reach", ctx, &funs);
+    let client = reach_invoke::Client::new("http://localhost:8080/reach", ctx, &funs);
     client.mail_pwd_send(&mail, &content, "测试", &()).await?;
     // wait for send
     tokio::time::sleep(Duration::from_secs(10)).await;
